backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.0.1
playwright==1.40.0
html2text==2024.2.26
openai==1.101.0
anthropic==0.64.0
httpx==0.27.2
psycopg2-binary==2.9.9
types-psycopg2==2.9.21.20
alembic==1.16.5
aiohttp==3.12.15
boto3==1.40.19
python-magic==0.4.27
pdfplumber==0.11.7
requests==2.32.5
types-requests==2.32.4.20250809
python-multipart==0.0.20

# Authentication dependencies
google-auth==2.40.3
google-auth-oauthlib==1.2.2

# Development dependencies
black==24.4.2
isort==5.13.2
flake8==7.1.1
flake8-annotations==3.1.1
mypy==1.11.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
